# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /registrations:
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Register for the conference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationCreated"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listRegistrations
      tags: [registrations]
      summary: List all registrations (admin)
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
        "401":
          description: Admin sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registrations/{registrationId}:
    delete:
      operationId: deleteRegistration
      tags: [registrations]
      summary: Delete a registration (admin)
      parameters:
        - $ref: "#/components/parameters/RegistrationId"
      responses:
        "204":
          description: Registration deleted
        "400":
          description: Invalid registration ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registrations/badge-request:
    post:
      operationId: requestExistingRegistrationBadge
      tags: [badges]
      summary: Request an attendee badge for an existing registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExistingRegistrationBadgeInput"
      responses:
        "200":
          description: Short-lived authorization for creating an attendee badge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeAccess"
        "400":
          description: Invalid badge request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Matching registration was not found or is not eligible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registrations/{registrationId}/badge/upload-url:
    post:
      operationId: requestRegistrationBadgeUploadUrl
      tags: [badges]
      summary: Request a private attendee photo upload URL
      parameters:
        - $ref: "#/components/parameters/RegistrationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BadgeUploadInput"
      responses:
        "200":
          description: Short-lived URL for uploading the attendee portrait
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgeUploadUrl"
        "400":
          description: Invalid badge upload request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired badge authorization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registrations/{registrationId}/badge/complete:
    post:
      operationId: completeRegistrationBadge
      tags: [badges]
      summary: Create and deliver an attendee badge
      parameters:
        - $ref: "#/components/parameters/RegistrationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BadgeCompleteInput"
      responses:
        "200":
          description: Badge delivered in the confirmation email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "202":
          description: Registration was confirmed but badge generation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Invalid badge completion request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired badge authorization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registrations/{registrationId}/badge/skip:
    post:
      operationId: skipRegistrationBadge
      tags: [badges]
      summary: Send the standard confirmation without an attendee badge
      parameters:
        - $ref: "#/components/parameters/RegistrationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BadgeSkipInput"
      responses:
        "200":
          description: Standard confirmation delivered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "401":
          description: Invalid or expired badge authorization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Registration:
      type: object
      required: [id, firstName, lastName, email, conferenceYear, volunteer, createdAt]
      properties:
        id:             { type: integer }
        firstName:      { type: string }
        lastName:       { type: string }
        email:          { type: string }
        phone:          { type: ["string", "null"] }
        smsConsent:     { type: boolean }
        conferenceYear: { type: integer }
        volunteer:      { type: boolean }
        volunteerRole:  { type: ["string", "null"] }
        badgeDeliveryStatus:
          type: string
          enum: [delivered, fallback]
        createdAt:      { type: string, format: date-time }
    RegistrationCreated:
      allOf:
        - $ref: "#/components/schemas/Registration"
        - type: object
          properties:
            badgeUploadToken:
              type: string
    RegistrationInput:
      type: object
      required: [firstName, lastName, email, conferenceYear]
      properties:
        firstName:      { type: string, minLength: 1 }
        lastName:       { type: string, minLength: 1 }
        email:          { type: string, format: email }
        phone:          { type: string }
        smsConsent:     { type: boolean, description: Whether the attendee opted in to receive SMS updates. }
        conferenceYear: { type: integer }
        volunteer:      { type: boolean }
        volunteerRole:  { type: string }
        wantsAttendeeBadge:
          type: boolean
          description: Whether a portrait will be uploaded to create an attendee badge.
        plusOne:
          $ref: "#/components/schemas/PlusOneInput"
...
    PlusOneInput:
      type: object
      description: An optional additional attendee registered with the primary attendee.
      required: [firstName, lastName, email]
      properties:
        firstName: { type: string, minLength: 1 }
        lastName:  { type: string, minLength: 1 }
        email:     { type: string, format: email }
        phone:     { type: string }
    ExistingRegistrationBadgeInput:
      type: object
      required: [firstName, lastName, email, conferenceYear]
      properties:
        firstName:      { type: string, minLength: 1 }
        lastName:       { type: string, minLength: 1 }
        email:          { type: string, format: email }
        conferenceYear: { type: integer }
    BadgeAccess:
      type: object
      required: [registrationId, badgeUploadToken]
      properties:
        registrationId: { type: integer }
        badgeUploadToken: { type: string }
    BadgeUploadInput:
      type: object
      required: [token, name, size, contentType]
      properties:
        token:       { type: string, minLength: 32 }
        name:        { type: string, minLength: 1, maxLength: 200 }
        size:        { type: integer, minimum: 1, maximum: 5242880 }
        contentType: { type: string, enum: [image/jpeg, image/png, image/webp] }
    BadgeUploadUrl:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:  { type: string, format: uri }
        objectPath: { type: string }
    BadgeCompleteInput:
      type: object
      required: [token, objectPath]
      properties:
        token:      { type: string, minLength: 32 }
        objectPath: { type: string, minLength: 1 }
    BadgeSkipInput:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 32 }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
  parameters:
...
    RegistrationId:
      name: registrationId
      in: path
      required: true
      schema:
        type: integer